refactor : controller 폴더 삭제, SlideContoller 폴더 하위에 자식 컴포넌트 위치
