+++ src/components/common/InfinityCarousel/components/SlideController/_mixins.scss
@@ -1,0 +1,7 @@
+@mixin controller-container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: rgba(0, 0, 0, 0.4);
+  height: 2.4rem;
+}
