+++ src/components/common/InfinityCarousel/components/SlideController/components/AutoplayButton/index.tsx
@@ -4,6 +4,8 @@ import React from 'react';
 import PauseIcon from '@/images/pause.svg';
 import PlayIcon from '@/images/play.svg';
 
+import styles from './index.module.scss';
+
 export interface AutoPlayButtonProps {
   isPlaying: boolean;
   disabled: boolean;
@@ -12,8 +14,8 @@ export interface AutoPlayButtonProps {
 
 const AutoPlayButton = ({ isPlaying, disabled, handlePlayButtonClick }: AutoPlayButtonProps) => {
   return (
-    <button onClick={handlePlayButtonClick} disabled={disabled}>
-      <Image src={isPlaying ? PauseIcon : PlayIcon} width={18} alt="" />
+    <button onClick={handlePlayButtonClick} disabled={disabled} className={styles.button}>
+      <Image src={isPlaying ? PauseIcon : PlayIcon} height={10} alt="" />
     </button>
   );
 };
